+++ packages/daf-core/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/uport-project/daf/compare/v2.4.1...v2.5.0) (2020-03-13)
+
+### Bug Fixes
+
+- GQL return raw while validating message ([53573bc](https://github.com/uport-project/daf/commit/53573bc07cbaabc92c34baaa3096acd279aa985f))
+
 ## [2.3.19](https://github.com/uport-project/daf/compare/v2.3.18...v2.3.19) (2020-03-11)
 
 ### Bug Fixes