+++ packages/daf-did-comm/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/uport-project/daf/compare/v2.4.1...v2.5.0) (2020-03-13)
+
+**Note:** Version bump only for package daf-did-comm
+
+## [2.4.1](https://github.com/uport-project/daf/compare/v2.4.0...v2.4.1) (2020-03-13)
+
+### Bug Fixes
+
+- Remove debug info ([a7753dd](https://github.com/uport-project/daf/commit/a7753ddb1875f5c7d02d1ad4f24b21f4963df338))
+
 ## [2.3.19](https://github.com/uport-project/daf/compare/v2.3.18...v2.3.19) (2020-03-11)
 
 **Note:** Version bump only for package daf-did-comm